+++ com.templlo.service.reservation/src/main/java/com/templlo/service/reservation/domain/temp/TestController.java
@@ -1,4 +1,4 @@
-package com.templlo.service.reservation.domain.checker.controller;
+package com.templlo.service.reservation.domain.temp;
 
 import com.templlo.service.reservation.domain.reservation.controller.exception.ReservationStatusCode;
 import com.templlo.service.reservation.domain.reservation.controller.exception.ReservationException;
